+++ WindrosePlus/Scripts/modules/query.lua
@@ -198,7 +198,8 @@ function Query._collectAndWrite()
             crop_speed = Query._config.getCropSpeedMultiplier(),
             weight = Query._config.getWeightMultiplier(),
             inventory_size = Query._config.getInventorySizeMultiplier(),
-            points_per_level = Query._config.getPointsPerLevelMultiplier()
+            points_per_level = Query._config.getPointsPerLevelMultiplier(),
+            cooking_speed = Query._config.getCookingSpeedMultiplier()
         },
         timestamp = os.time()
     }